refactor(aws): move accounts out of the CredentialsConfig class * this will make it easier to custom bind accounts when loading them from a local config file. This is for the following issue: https://github.com/spinnaker/spinnaker/issues/6492
