+++ clouddriver-aws/src/main/java/com/netflix/spinnaker/clouddriver/aws/security/config/CredentialsConfig.java
@@ -18,10 +18,6 @@ package com.netflix.spinnaker.clouddriver.aws.security.config;
 
 import static lombok.EqualsAndHashCode.Include;
 
-import com.netflix.spinnaker.credentials.definition.CredentialsDefinition;
-import com.netflix.spinnaker.fiat.model.resources.Permissions;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import lombok.EqualsAndHashCode;
 
@@ -138,232 +134,6 @@ public class CredentialsConfig {
     }
   }
 
-  @EqualsAndHashCode(onlyExplicitlyIncluded = true)
-  public static class Account implements CredentialsDefinition {
-    @Include private String name;
-    @Include private String environment;
-    @Include private String accountType;
-    @Include private String accountId;
-    @Include private String defaultKeyPair;
-    @Include private Boolean enabled;
-    @Include private List<Region> regions;
-    @Include private List<String> defaultSecurityGroups;
-    private List<String> requiredGroupMembership;
-    @Include private Permissions.Builder permissions;
-    @Include private String edda;
-    @Include private Boolean eddaEnabled;
-    @Include private Boolean lambdaEnabled;
-    @Include private String discovery;
-    @Include private Boolean discoveryEnabled;
-    @Include private String front50;
-    @Include private Boolean front50Enabled;
-    @Include private String bastionHost;
-    @Include private Boolean bastionEnabled;
-    @Include private String assumeRole;
-    @Include private String sessionName;
-    @Include private String externalId;
-    @Include private List<LifecycleHook> lifecycleHooks;
-    @Include private boolean allowPrivateThirdPartyImages;
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-
-    public String getEnvironment() {
-      return environment;
-    }
-
-    public void setEnvironment(String environment) {
-      this.environment = environment;
-    }
-
-    public String getAccountType() {
-      return accountType;
-    }
-
-    public void setAccountType(String accountType) {
-      this.accountType = accountType;
-    }
-
-    public String getAccountId() {
-      return accountId;
-    }
-
-    public void setAccountId(String accountId) {
-      this.accountId = accountId;
-    }
-
-    public String getDefaultKeyPair() {
-      return defaultKeyPair;
-    }
-
-    public void setDefaultKeyPair(String defaultKeyPair) {
-      this.defaultKeyPair = defaultKeyPair;
-    }
-
-    public Boolean getEnabled() {
-      return enabled;
-    }
-
-    public void setEnabled(Boolean enabled) {
-      this.enabled = enabled;
-    }
-
-    public List<Region> getRegions() {
-      return regions;
-    }
-
-    public void setRegions(List<Region> regions) {
-      if (regions != null) {
-        regions.sort(Comparator.comparing(Region::getName));
-      }
-      this.regions = regions;
-    }
-
-    public List<String> getDefaultSecurityGroups() {
-      return defaultSecurityGroups;
-    }
-
-    public void setDefaultSecurityGroups(List<String> defaultSecurityGroups) {
-      if (defaultSecurityGroups != null) {
-        Collections.sort(defaultSecurityGroups);
-      }
-      this.defaultSecurityGroups = defaultSecurityGroups;
-    }
-
-    public List<String> getRequiredGroupMembership() {
-      return requiredGroupMembership;
-    }
-
-    public void setRequiredGroupMembership(List<String> requiredGroupMembership) {
-      this.requiredGroupMembership = requiredGroupMembership;
-    }
-
-    public Permissions.Builder getPermissions() {
-      return permissions;
-    }
-
-    public void setPermissions(Permissions.Builder permissions) {
-      this.permissions = permissions;
-    }
-
-    public String getEdda() {
-      return edda;
-    }
-
-    public void setEdda(String edda) {
-      this.edda = edda;
-    }
-
-    public Boolean getEddaEnabled() {
-      return eddaEnabled;
-    }
-
-    public void setEddaEnabled(Boolean eddaEnabled) {
-      this.eddaEnabled = eddaEnabled;
-    }
-
-    public String getDiscovery() {
-      return discovery;
-    }
-
-    public void setDiscovery(String discovery) {
-      this.discovery = discovery;
-    }
-
-    public Boolean getDiscoveryEnabled() {
-      return discoveryEnabled;
-    }
-
-    public void setDiscoveryEnabled(Boolean discoveryEnabled) {
-      this.discoveryEnabled = discoveryEnabled;
-    }
-
-    public String getFront50() {
-      return front50;
-    }
-
-    public void setFront50(String front50) {
-      this.front50 = front50;
-    }
-
-    public Boolean getFront50Enabled() {
-      return front50Enabled;
-    }
-
-    public void setFront50Enabled(Boolean front50Enabled) {
-      this.front50Enabled = front50Enabled;
-    }
-
-    public String getBastionHost() {
-      return bastionHost;
-    }
-
-    public void setBastionHost(String bastionHost) {
-      this.bastionHost = bastionHost;
-    }
-
-    public Boolean getBastionEnabled() {
-      return bastionEnabled;
-    }
-
-    public void setBastionEnabled(Boolean bastionEnabled) {
-      this.bastionEnabled = bastionEnabled;
-    }
-
-    public String getAssumeRole() {
-      return assumeRole;
-    }
-
-    public void setAssumeRole(String assumeRole) {
-      this.assumeRole = assumeRole;
-    }
-
-    public String getSessionName() {
-      return sessionName;
-    }
-
-    public void setSessionName(String sessionName) {
-      this.sessionName = sessionName;
-    }
-
-    public String getExternalId() {
-      return externalId;
-    }
-
-    public void setExternalId(String externalId) {
-      this.externalId = externalId;
-    }
-
-    public List<LifecycleHook> getLifecycleHooks() {
-      return lifecycleHooks;
-    }
-
-    public void setLifecycleHooks(List<LifecycleHook> lifecycleHooks) {
-      this.lifecycleHooks = lifecycleHooks;
-    }
-
-    public Boolean getAllowPrivateThirdPartyImages() {
-      return allowPrivateThirdPartyImages;
-    }
-
-    public void setAllowPrivateThirdPartyImages(Boolean allowPrivateThirdPartyImages) {
-      this.allowPrivateThirdPartyImages = allowPrivateThirdPartyImages;
-    }
-
-    public Boolean getLambdaEnabled() {
-      return lambdaEnabled;
-    }
-
-    public void setLambdaEnabled(Boolean lambdaEnabled) {
-      this.lambdaEnabled = lambdaEnabled;
-    }
-  }
-
   private String accessKeyId;
   private String secretAccessKey;
   private String defaultKeyPairTemplate;
@@ -378,8 +148,6 @@ public class CredentialsConfig {
   private String defaultSessionName;
   private String defaultLifecycleHookRoleARNTemplate;
   private String defaultLifecycleHookNotificationTargetARNTemplate;
-
-  private List<Account> accounts;
 
   public String getDefaultKeyPairTemplate() {
     return defaultKeyPairTemplate;
@@ -451,14 +219,6 @@ public class CredentialsConfig {
 
   public void setDefaultSessionName(String defaultSessionName) {
     this.defaultSessionName = defaultSessionName;
-  }
-
-  public List<Account> getAccounts() {
-    return accounts;
-  }
-
-  public void setAccounts(List<Account> accounts) {
-    this.accounts = accounts;
   }
 
   public List<LifecycleHook> getDefaultLifecycleHooks() {
